backend/requirements.txt
Flask
flask-restful
requests
google-generativeai
python-Levenshtein
pillow 
fuzzywuzzy
flask-cors